Refactor to use components
